expose stash/unstash row & column functions
